requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.5
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
isodate==0.6.1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0
